requirements.txt
python-dotenv
openai
langchain-openai
langchain-ollama
langchain
langchain-community
langchain-core
langchain-cohere
ollama
yfinance
streamlit
pypdf
PyPDF2
pymupdf
tiktoken
langchain
langgraph
faiss-cpu
pytesseract
pillow
Document
gtts